refactor: extract Campaign and Form fields to a shared component
