+++ src/Admin/fields/CampaignForm/types.ts
@@ -1,0 +1,13 @@
+export interface CampaignFormProps {
+    campaignId: number;
+    formId: number;
+    campaignsWithForms: {
+        [campaignId: string]: {
+            title: string;
+            defaultFormId: string;
+            forms: {
+                [formId: string]: string;
+            };
+        };
+    };
+}
